api/requirements.txt
flask==2.1.0
flask-cors==3.0.10
prometheus-client==0.13.1
kafka-python==2.0.2
werkzeug==2.1.1
uuid

